requirements.txt
Django==1.5.1
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.31
celery==3.0.21
celery-with-redis==3.0
celerymon==1.0.3
django-celery==3.0.21
django-sse==0.4.1
gevent==0.13.8
greenlet==0.4.1
gunicorn==17.5
kombu==2.5.12
python-dateutil==2.1
pytz==2013b
redis==2.7.6
six==1.3.0
sse==1.2
tornado==3.1
wsgiref==0.1.2
Markdown==2.3.1
South==0.8.2
Werkzeug==0.9.4
django-extensions==1.2.1
djangorestframework==2.3.8
ipython==1.1.0
